Add blade support
