+++ composer.json
@@ -32,7 +32,9 @@
         "phpstan/phpstan": "^1.10.0",
         "phpunit/phpunit": "^10.2",
         "spaze/phpstan-disallowed-calls": "^3.1",
-        "symplify/monorepo-builder": "^11.2"
+        "symplify/monorepo-builder": "^11.2",
+        "jenssegers/blade": "^2.0",
+        "illuminate/view": "11.7.0"
     },
     "autoload": {
         "files": [
