+++ src/Tempest/View/Renderers/BladeConfig.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tempest\View\Renderers;
+
+final readonly class BladeConfig
+{
+    public function __construct(
+        public array $viewPaths = [],
+        public ?string $cachePath = null,
+    ) {}
+}
