+++ tests/Integration/View/blade/base.blade.php
@@ -1,0 +1,3 @@
+<html>
+{{ $slot }}
+</html>
