+++ tests/Integration/View/blade/cache/83bb9a9d9ceab93bd9480c0106d1f8cc.php
@@ -1,0 +1,4 @@
+<html>
+<?php echo e($slot); ?>
+
+</html><?php /**PATH /Users/brentroose/Dev/tempest-framework/tests/Integration/View/blade/base.blade.php ENDPATH**/ ?>
