+++ tests/Integration/View/blade/cache/f7fa96413bddc8e698fe821def33c5f7.php
@@ -1,0 +1,3 @@
+<?php $__env->startComponent('base'); ?>
+    Hi
+<?php echo $__env->renderComponent(); ?><?php /**PATH /Users/brentroose/Dev/tempest-framework/tests/Integration/View/blade/index.blade.php ENDPATH**/ ?>
