+++ tests/Integration/View/blade/index.blade.php
@@ -1,0 +1,3 @@
+@component('base')
+    Hi
+@endcomponent
